+++ src/main/java/pl/edu/pw/mini/ingreedio/api/review/repository/ReviewRepository.java
@@ -10,10 +10,10 @@ import pl.edu.pw.mini.ingreedio.api.review.model.Review;
 
 @Repository
 public interface ReviewRepository extends JpaRepository<Review, Long> {
-    Optional<Review> findById(Long id);
+    Optional<Review> findById(int id);
 
     @Query("SELECT r FROM Review r WHERE r.user.id = :userId AND r.productId = :productId")
-    Optional<Review> findByUserIdAndProductId(@Param("userId") Long userId,
+    Optional<Review> findByUserIdAndProductId(@Param("userId") int userId,
                                               @Param("productId") Long productId);
 
     @Query("SELECT r FROM Review r WHERE r.productId = :productId")
